+++ lib/xmat-models/xmat-file.model.ts
@@ -1,0 +1,5 @@
+export interface XmatFile {
+    desc?: string;
+    fileName: string;
+    url: string;
+}